requirements.txt
tensorflow==2.19.1
fastapi
uvicorn
python-multipart
pillow
tensorflow-serving-api==2.9.0
matplotlib
numpy